requirements.txt
numpy
torch 
opencv-python-headless 
pyyaml 
matplotlib
tqdm
pillow